Geolocation added

+++ app/src/main/java/com/example/weather/weather_model/Current.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Current(val sunrise: Int = 0,
                    val temp: Double = 0.0,
                    val visibility: Int = 0,
-                   val uvi: Int = 0,
+                   val uvi: Double = 0.0,
                    val pressure: Int = 0,
                    val clouds: Int = 0,
                    @SerializedName("feels_like") val feelsLike: Double = 0.0,
